+++ nac_test/utils/__init__.py
@@ -12,7 +12,7 @@ from nac_test.utils.device_validation import (
 )
 from nac_test.utils.environment import EnvironmentValidator
 from nac_test.utils.file_discovery import find_data_file
-from nac_test.utils.logging import VerbosityLevel, configure_logging
+from nac_test.utils.logging import LogLevel, configure_logging
 from nac_test.utils.strings import sanitize_hostname
 from nac_test.utils.system_resources import SystemResourceCalculator
 from nac_test.utils.terminal import terminal
@@ -24,7 +24,7 @@ __all__ = [
     "cleanup_pyats_runtime",
     "cleanup_old_test_outputs",
     "configure_logging",
-    "VerbosityLevel",
+    "LogLevel",
     # Asyncio utilities (cross-version compatibility)
     "get_or_create_event_loop",
     # Device validation utilities (SSH/D2D architecture)